+++ ProtoActor.sln
@@ -61,6 +61,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "ReceiveTimeout", "ReceiveTi
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ReceiveTimeout", "examples\ReceiveTimeout\ReceiveTimeout.csproj", "{AC11D152-F1FD-4A48-A0DC-0B315D507457}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Proto.Mailbox", "src\Proto.Mailbox\Proto.Mailbox.csproj", "{7B1E82F7-270F-40F0-AF76-3FA54EF79124}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -275,6 +277,18 @@ Global
 		{AC11D152-F1FD-4A48-A0DC-0B315D507457}.Release|x64.Build.0 = Release|x64
 		{AC11D152-F1FD-4A48-A0DC-0B315D507457}.Release|x86.ActiveCfg = Release|x86
 		{AC11D152-F1FD-4A48-A0DC-0B315D507457}.Release|x86.Build.0 = Release|x86
+		{7B1E82F7-270F-40F0-AF76-3FA54EF79124}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7B1E82F7-270F-40F0-AF76-3FA54EF79124}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7B1E82F7-270F-40F0-AF76-3FA54EF79124}.Debug|x64.ActiveCfg = Debug|x64
+		{7B1E82F7-270F-40F0-AF76-3FA54EF79124}.Debug|x64.Build.0 = Debug|x64
+		{7B1E82F7-270F-40F0-AF76-3FA54EF79124}.Debug|x86.ActiveCfg = Debug|x86
+		{7B1E82F7-270F-40F0-AF76-3FA54EF79124}.Debug|x86.Build.0 = Debug|x86
+		{7B1E82F7-270F-40F0-AF76-3FA54EF79124}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7B1E82F7-270F-40F0-AF76-3FA54EF79124}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7B1E82F7-270F-40F0-AF76-3FA54EF79124}.Release|x64.ActiveCfg = Release|x64
+		{7B1E82F7-270F-40F0-AF76-3FA54EF79124}.Release|x64.Build.0 = Release|x64
+		{7B1E82F7-270F-40F0-AF76-3FA54EF79124}.Release|x86.ActiveCfg = Release|x86
+		{7B1E82F7-270F-40F0-AF76-3FA54EF79124}.Release|x86.Build.0 = Release|x86
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -306,5 +320,6 @@ Global
 		{2C02A779-3805-42D3-A8F6-12940E6BDF34} = {768E2FB6-A663-4D6A-A791-B2E1AF7296AF}
 		{C78E50DC-8D65-4117-9B3E-BFA3F4A8101E} = {59DCCC96-DDAF-469F-9E8E-9BC733285082}
 		{AC11D152-F1FD-4A48-A0DC-0B315D507457} = {C78E50DC-8D65-4117-9B3E-BFA3F4A8101E}
+		{7B1E82F7-270F-40F0-AF76-3FA54EF79124} = {771514F1-12AE-4A26-89CB-2646D3EF7034}
 	EndGlobalSection
 EndGlobal
