+++ src/Proto.Actor/Context.cs
@@ -12,14 +12,6 @@ using System.Threading.Tasks;
 
 namespace Proto
 {
-    public interface IMessageInvoker
-    {
-        Task InvokeSystemMessageAsync(object msg);
-        Task InvokeUserMessageAsync(object msg);
-
-        void EscalateFailure(Exception reason, object message);
-    }
-
     public interface IContext
     {
         PID Parent { get; }
