+++ src/Proto.Mailbox/BoundedMailboxQueue.cs
@@ -1,0 +1,26 @@
+﻿namespace Proto
+{
+    public class BoundedMailboxQueue : IMailboxQueue
+    {
+        private readonly MPMCQueue _messages;
+
+        public BoundedMailboxQueue(int size)
+        {
+            _messages = new MPMCQueue(size);
+        }
+
+        public void Push(object message)
+        {
+            _messages.Enqueue(message);
+        }
+
+        public object Pop()
+        {
+            return _messages.TryDequeue(out var message)
+                ? message
+                : null;
+        }
+
+        public bool HasMessages => _messages.Count > 0;
+    }
+}
