+++ src/Proto.Mailbox/Queue.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Proto
+{
+
+    public interface IMailboxQueue
+    {
+        bool HasMessages { get; }
+        void Push(object message);
+        object Pop();
+    }
+}
